+++ libraries/ml/src/MCL/clustering-view.ts
@@ -8,7 +8,6 @@ import {KnownMetrics} from '../typed-metrics';
 import {DistanceAggregationMethod} from '../distance-matrix/types';
 import {PreprocessFunctionReturnType} from '../functionEditors/dimensionality-reduction-editor';
 import {Options} from '@datagrok-libraries/utils/src/type-declarations';
-import { SparseMatrix } from '@datagrok-libraries/math';
 
 
 export type MCLClusterViewerResult = {
@@ -24,8 +23,7 @@ export async function markovCluster(
   df: DG.DataFrame, cols: DG.Column[], metrics: KnownMetrics[],
   weights: number[], aggregationMethod: DistanceAggregationMethod, preprocessingFuncs: (DG.Func | null | undefined)[],
   preprocessingFuncArgs: any[], threshold: number = 80, maxIterations: number = 10,
-  useWebGPU: boolean = false, inflate: number = 2, minClusterSize: number = 5, scp?: DG.ScatterPlotViewer,
-  presetMatrix?: SparseMatrix // TODO: DONT FORGET TO REMOVE THIS PARAMETER
+  useWebGPU: boolean = false, inflate: number = 2, minClusterSize: number = 5, scp?: DG.ScatterPlotViewer
 ): Promise<undefined | MCLClusterViewerResult> {
   const scatterPlotProps = {
     showXAxis: false,
@@ -59,7 +57,7 @@ export async function markovCluster(
   }
 
   const mclWorker = createMCLWorker(encodedColEntries.map((it) => it.entries),
-    threshold, weights, aggregationMethod, metrics, distanceFnArgs, maxIterations, useWebGPU, inflate, presetMatrix);
+    threshold, weights, aggregationMethod, metrics, distanceFnArgs, maxIterations, useWebGPU, inflate);
 
   const terminateSub = grok.events.onViewerClosed.subscribe((args) => {
     if (args.args.viewer?.props?.title === sc.props.title && sc.type === args.args?.viewer?.type) {